requirements.txt
paho-mqtt==2.1.0
pyserial==3.5
PyYAML==6.0.1
typing_extensions==4.11.0
